updated CRUD operations and other helper functions
